chore(ci): split slow source tests into parallel files
Extracts world and US source validation into separate test files so tap
can run them in parallel with --jobs 4. Shared validation logic moved
to test/source-helpers.js. Adds geometry result cache to avoid
re-validating the same GeoJSON file across multiple version configs.
Also fixes glob pattern to skip generated topo.json files in sources/.
